requirements.txt
fastapi==0.115.11
fastapi-cli==0.0.7
PyJWT==2.10.1
passlib==1.7.4
bcrypt==4.3.0
psycopg2-binary==2.9.10
python-multipart==0.0.20
sqlmodel==0.0.24
pgvector==0.4.0
sentence-transformers==3.4.1
openai==1.66.3
oso-cloud==2.4.2
asyncpg==0.30.0
greenlet==3.1.1
